Hii-17: unstagged changes.
